+++ client/src/store/reducers/header/categories/CategoriesSlice.js
@@ -2,7 +2,7 @@ import {createSlice} from "@reduxjs/toolkit";
 
 const initialState = {
     categories: [],
-    selectedCategory: {},
+    selectedCategory: {}
 }
 
 export const categoriesSlice = createSlice({
@@ -12,9 +12,9 @@ export const categoriesSlice = createSlice({
         setCategories(state, action) {
            state.categories = action.payload
         },
-        setSelectedCategory(state, action) {
-           state.selectedCategory = action.payload
-        },
+        setSelectedCategory(state,action) {
+            state.selectedCategory = action.payload
+        }
     }
 })
